improve menu popover height restriction
by re-using PopoverContentContext to avoid complicated requestAnimationFrame
and other logic based on "open" state

also directly set minHeight on element.style instead of relying on state which
can get stale
